use old benchmark file
